refactor: 가격 계산 응답 DTO 필드명 수정
