+++ service/catalog/src/main/java/com/bappul/catalog/application/mapper/PricingMapper.java
@@ -3,7 +3,7 @@ package com.bappul.catalog.application.mapper;
 import com.bappul.catalog.domain.entity.Menu;
 import com.bappul.catalog.domain.entity.Store;
 import com.bappul.catalog.web.v1.response.menu.internal.CartItemCalculateResponse;
-import com.bappul.catalog.web.v1.response.menu.internal.OptionPerPrice;
+import com.bappul.catalog.web.v1.response.menu.internal.OptionPrice;
 import com.bappul.catalog.web.v1.response.menu.internal.PricingInternalResponse;
 import java.math.BigDecimal;
 import java.util.List;
@@ -17,7 +17,7 @@ public interface PricingMapper {
   @Mapping(target = "menuName", source = "menu.name")
   CartItemCalculateResponse toCartItemCalculateResponse(Menu menu,
       BigDecimal basePrice,
-      List<OptionPerPrice> optionPerPrices,
+      List<OptionPrice> optionPrices,
       BigDecimal optionUnitPrice,
       BigDecimal unitPrice,
       int quantity,
@@ -26,6 +26,7 @@ public interface PricingMapper {
 
   @Mapping(target = "storeId", source = "store.id")
   @Mapping(target = "storeName", source = "store.name")
+  @Mapping(target = "deliveryFeePrice", source = "store.deliveryFee")
   PricingInternalResponse toPricingInternalResponse(
       Store store,
       List<CartItemCalculateResponse> items,
