+++ service/catalog/src/main/java/com/bappul/catalog/web/v1/response/menu/internal/PricingInternalResponse.java
@@ -16,5 +16,6 @@ public class PricingInternalResponse {
   Long storeId;
   String storeName;
   List<CartItemCalculateResponse> items;
+  BigDecimal deliveryFeePrice;
   BigDecimal totalPrice;
 }
