+++ service/order/src/main/java/com/bappul/order/adapter/response/CartItemCalculateResponse.java
@@ -1,4 +1,4 @@
-package com.bappul.catalog.web.v1.response.menu.internal;
+package com.bappul.order.adapter.response;
 
 import java.math.BigDecimal;
 import java.util.List;
